Write a new description for gamepad in order to host understand that device as gamepad nor mouse. Also add some function to descriptor about Force Feedback

+++ Core/Src/Joystick ADC.c
@@ -21,8 +21,8 @@ void ADC_DMA_Init (void)
 
 void ScaleData (void)
 {
-	adc_x1 = (int8_t)(((float)adc_buffer1 [0] / 4095.0) * 510) - 255;
-	adc_y1 = (int8_t)(((float)adc_buffer1 [1] / 4095.0) * 510) - 255;
-	adc_x2 = (int8_t)(((float)adc_buffer2 [0] / 4095.0) * 510) - 255;
-	adc_y2 = (int8_t)(((float)adc_buffer2 [1] / 4095.0) * 510) - 255;
+	adc_x1 = (int8_t)((adc_buffer1 [0] / 4095) * 510) - 255;
+	adc_y1 = (int8_t)((adc_buffer1 [1] / 4095) * 510) - 255;
+	adc_x2 = (int8_t)((adc_buffer2 [0] / 4095) * 510) - 255;
+	adc_y2 = (int8_t)((adc_buffer2 [1] / 4095) * 510) - 255;
 }
